TFIP-311 - implemented validation of time spent and auto formating of timespant time - implemented date range picker - implemented couting of time spent sum - implemented edit dialog - implemented displaying author select and author column in table olny when user with admin role is logged in - formatted dates and time for tables and fixed sorting of dates in table - added confirm dialog for deleting work log

+++ resources/pages/office-absences.vue
@@ -278,7 +278,11 @@
       },
       async deleteItem(item) {
         if (item.absenceState.name === 'WAITING_FOR_APPROVAL') {
-          await this.$store.dispatch('deleteOfficeAbsence', item.id);
+          const confirmed = confirm(`Opravdu chcete smazat tuhle žádost o nepřítomnost.`);
+
+          if (confirmed) {
+            await this.$store.dispatch('deleteOfficeAbsence', item.id);
+          }
         } else {
           this.$refs.cancelAbsenceDialog.openDialog(item);
         }
